deps: drop OTel from direct requires; add -tags=otel CI lane
- Move go.opentelemetry.io/otel/metric to indirect (no direct user left).
- Gate examples/herd-hub-leaf telemetry behind -tags=otel: setupTelemetry
  and pingOTLPCollector live in telemetry_otel.go; the default build gets
  a no-op stub from telemetry_default.go. main.go drops all direct OTel
  imports so default `go build ./...` no longer compiles otel/sdk and
  otel/exporters into the herd binary.
- Remove coord/otel_recorder_test.go: dead helper (no callers in tree),
  was the last otel/sdk import outside the telemetry seam.
- Add CI lane: build + test under -tags=otel so the otel-tagged paths
  cannot bit-rot. Default build still goes through `make check`.

Net effect: bones-owned source has zero direct OTel imports under
default tags. The remaining 25 transitive otel deps in
`go list -deps ./cmd/bones` come from EdgeSync's otelhttp client; the
seam in internal/telemetry collapses to a no-op without them. Stripped
binary unchanged at 24M (otelhttp's transitive set already covers what
the seam needs under -tags=otel).

Audit ref: docs/code-review/2026-04-28-hipp-audit.md §2
Ousterhout ref: docs/code-review/2026-04-28-ousterhout-plan-audit.md §Phase 2
