+++ .github/workflows/ci.yml
@@ -52,3 +52,15 @@ jobs:
       - name: make check
         working-directory: bones
         run: make check
+
+      # Phase 2 of the Hipp audit gates direct OTel use behind -tags=otel.
+      # Default builds must stay OTel-free (the seam in internal/telemetry
+      # collapses to a no-op); this lane proves the otel-tagged paths still
+      # compile and pass tests so the seam can't bit-rot.
+      - name: build (-tags=otel)
+        working-directory: bones
+        run: go build -tags=otel ./...
+
+      - name: test (-tags=otel)
+        working-directory: bones
+        run: go test -tags=otel ./...
